+++ lib/container.php
@@ -38,9 +38,8 @@ Final class Container implements Container_Interface {
 	 *
 	 * @return Container
 	 */
-	public static function getInstance()
-	{
-		if( null == static::$instance ){
+	public static function getInstance() {
+		if ( null == static::$instance ) {
 			static::$instance = new static();
 		}