+++ model/estate.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace wp_whise\model;
+
+class Estate {
+
+
+}