+++ model/whise-interface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace wp_whise\model;
-
-interface Whise_Interface {
-
-	public function get_houses($page);
-}