+++ model/whise.php
@@ -1,7 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: aytackokus
- * Date: 04/07/2018
- * Time: 16:41
- */