+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # CHANGELOG WP-Whise
 
 ## v1.0.0 - 04/07/2018
+- Whise adapter & controller & unit tests
+- Register custom post types